chore/remove payouts from server components (#1621)
Follow up of a PR review in
https://github.com/antiwork/gumroad/pull/1609#discussion_r2419657071

This pull request removes the `BalancePage` (Payouts) component from the
list of SSR components powered by React on Rails, since we no longer use
React on Rails but Inertia to render it.

It also:
- Renames the inertia template from `index` to `Index` to match
Inertia's conventions and avoid confusion with how `index` files are
used
- Moves the BalancePage template from `app/components/server-components`
to `app/components` for consistency
- Renames the `BalancePage` component to `Payouts` to be consistent with
the page name
- Identified code smells and added TODO comments

### AI Disclosure

No AI was used at all.

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>